KellyErrorEstimator: change output vector datatype to double.

+++ include/deal.II/numerics/error_estimator.h
@@ -345,7 +345,7 @@ public:
                    const Function<spacedim, typename InputVector::value_type> *>
       &                       neumann_bc,
     const InputVector &       solution,
-    Vector<float> &           error,
+    Vector<double> &          error,
     const ComponentMask &     component_mask = ComponentMask(),
     const Function<spacedim> *coefficients   = nullptr,
     const unsigned int        n_threads      = numbers::invalid_unsigned_int,
@@ -366,7 +366,7 @@ public:
                    const Function<spacedim, typename InputVector::value_type> *>
       &                       neumann_bc,
     const InputVector &       solution,
-    Vector<float> &           error,
+    Vector<double> &          error,
     const ComponentMask &     component_mask = ComponentMask(),
     const Function<spacedim> *coefficients   = nullptr,
     const unsigned int        n_threads      = numbers::invalid_unsigned_int,
@@ -397,7 +397,7 @@ public:
                    const Function<spacedim, typename InputVector::value_type> *>
       &                                     neumann_bc,
     const std::vector<const InputVector *> &solutions,
-    std::vector<Vector<float> *> &          errors,
+    std::vector<Vector<double> *> &         errors,
     const ComponentMask &                   component_mask = ComponentMask(),
     const Function<spacedim> *              coefficients   = nullptr,
     const unsigned int        n_threads    = numbers::invalid_unsigned_int,
@@ -418,7 +418,7 @@ public:
                    const Function<spacedim, typename InputVector::value_type> *>
       &                                     neumann_bc,
     const std::vector<const InputVector *> &solutions,
-    std::vector<Vector<float> *> &          errors,
+    std::vector<Vector<double> *> &         errors,
     const ComponentMask &                   component_mask = ComponentMask(),
     const Function<spacedim> *              coefficients   = nullptr,
     const unsigned int        n_threads    = numbers::invalid_unsigned_int,
@@ -441,14 +441,13 @@ public:
                    const Function<spacedim, typename InputVector::value_type> *>
       &                       neumann_bc,
     const InputVector &       solution,
-    Vector<float> &           error,
+    Vector<double> &          error,
     const ComponentMask &     component_mask = ComponentMask(),
     const Function<spacedim> *coefficients   = nullptr,
     const unsigned int        n_threads      = numbers::invalid_unsigned_int,
     const types::subdomain_id subdomain_id   = numbers::invalid_subdomain_id,
     const types::material_id  material_id    = numbers::invalid_material_id,
     const Strategy            strategy       = cell_diameter_over_24);
-
 
   /**
    * Equivalent to the set of functions above, except that this one takes a
@@ -463,7 +462,7 @@ public:
                    const Function<spacedim, typename InputVector::value_type> *>
       &                       neumann_bc,
     const InputVector &       solution,
-    Vector<float> &           error,
+    Vector<double> &          error,
     const ComponentMask &     component_mask = ComponentMask(),
     const Function<spacedim> *coefficients   = nullptr,
     const unsigned int        n_threads      = numbers::invalid_unsigned_int,
@@ -486,7 +485,7 @@ public:
                    const Function<spacedim, typename InputVector::value_type> *>
       &                                     neumann_bc,
     const std::vector<const InputVector *> &solutions,
-    std::vector<Vector<float> *> &          errors,
+    std::vector<Vector<double> *> &         errors,
     const ComponentMask &                   component_mask = ComponentMask(),
     const Function<spacedim> *              coefficients   = nullptr,
     const unsigned int        n_threads    = numbers::invalid_unsigned_int,
@@ -508,7 +507,7 @@ public:
                    const Function<spacedim, typename InputVector::value_type> *>
       &                                     neumann_bc,
     const std::vector<const InputVector *> &solutions,
-    std::vector<Vector<float> *> &          errors,
+    std::vector<Vector<double> *> &         errors,
     const ComponentMask &                   component_mask = ComponentMask(),
     const Function<spacedim> *              coefficients   = nullptr,
     const unsigned int        n_threads    = numbers::invalid_unsigned_int,
@@ -633,7 +632,7 @@ public:
                    const Function<spacedim, typename InputVector::value_type> *>
       &                       neumann_bc,
     const InputVector &       solution,
-    Vector<float> &           error,
+    Vector<double> &          error,
     const ComponentMask &     component_mask = ComponentMask(),
     const Function<spacedim> *coefficient    = nullptr,
     const unsigned int        n_threads      = numbers::invalid_unsigned_int,
@@ -656,7 +655,7 @@ public:
                    const Function<spacedim, typename InputVector::value_type> *>
       &                       neumann_bc,
     const InputVector &       solution,
-    Vector<float> &           error,
+    Vector<double> &          error,
     const ComponentMask &     component_mask = ComponentMask(),
     const Function<spacedim> *coefficients   = nullptr,
     const unsigned int        n_threads      = numbers::invalid_unsigned_int,
@@ -689,7 +688,7 @@ public:
                    const Function<spacedim, typename InputVector::value_type> *>
       &                                     neumann_bc,
     const std::vector<const InputVector *> &solutions,
-    std::vector<Vector<float> *> &          errors,
+    std::vector<Vector<double> *> &         errors,
     const ComponentMask &                   component_mask = ComponentMask(),
     const Function<spacedim> *              coefficients   = nullptr,
     const unsigned int        n_threads    = numbers::invalid_unsigned_int,
@@ -712,7 +711,7 @@ public:
                    const Function<spacedim, typename InputVector::value_type> *>
       &                                     neumann_bc,
     const std::vector<const InputVector *> &solutions,
-    std::vector<Vector<float> *> &          errors,
+    std::vector<Vector<double> *> &         errors,
     const ComponentMask &                   component_mask = ComponentMask(),
     const Function<spacedim> *              coefficients   = nullptr,
     const unsigned int        n_threads    = numbers::invalid_unsigned_int,
@@ -736,7 +735,7 @@ public:
                    const Function<spacedim, typename InputVector::value_type> *>
       &                       neumann_bc,
     const InputVector &       solution,
-    Vector<float> &           error,
+    Vector<double> &          error,
     const ComponentMask &     component_mask = ComponentMask(),
     const Function<spacedim> *coefficients   = nullptr,
     const unsigned int        n_threads      = numbers::invalid_unsigned_int,
@@ -759,7 +758,7 @@ public:
                    const Function<spacedim, typename InputVector::value_type> *>
       &                       neumann_bc,
     const InputVector &       solution,
-    Vector<float> &           error,
+    Vector<double> &          error,
     const ComponentMask &     component_mask = ComponentMask(),
     const Function<spacedim> *coefficients   = nullptr,
     const unsigned int        n_threads      = numbers::invalid_unsigned_int,
@@ -783,7 +782,7 @@ public:
                    const Function<spacedim, typename InputVector::value_type> *>
       &                                     neumann_bc,
     const std::vector<const InputVector *> &solutions,
-    std::vector<Vector<float> *> &          errors,
+    std::vector<Vector<double> *> &         errors,
     const ComponentMask &                   component_mask = ComponentMask(),
     const Function<spacedim> *              coefficients   = nullptr,
     const unsigned int        n_threads    = numbers::invalid_unsigned_int,
@@ -806,7 +805,7 @@ public:
                    const Function<spacedim, typename InputVector::value_type> *>
       &                                     neumann_bc,
     const std::vector<const InputVector *> &solutions,
-    std::vector<Vector<float> *> &          errors,
+    std::vector<Vector<double> *> &         errors,
     const ComponentMask &                   component_mask = ComponentMask(),
     const Function<spacedim> *              coefficients   = nullptr,
     const unsigned int        n_threads    = numbers::invalid_unsigned_int,
