+++ source/numerics/error_estimator_1d.cc
@@ -66,7 +66,7 @@ KellyErrorEstimator<1, spacedim>::estimate(
                  const Function<spacedim, typename InputVector::value_type> *>
     &                       neumann_bc,
   const InputVector &       solution,
-  Vector<float> &           error,
+  Vector<double> &          error,
   const ComponentMask &     component_mask,
   const Function<spacedim> *coefficients,
   const unsigned int        n_threads,
@@ -76,7 +76,7 @@ KellyErrorEstimator<1, spacedim>::estimate(
 {
   // just pass on to the other function
   const std::vector<const InputVector *> solutions(1, &solution);
-  std::vector<Vector<float> *>           errors(1, &error);
+  std::vector<Vector<double> *>          errors(1, &error);
   estimate(mapping,
            dof_handler,
            quadrature,
@@ -103,7 +103,7 @@ KellyErrorEstimator<1, spacedim>::estimate(
                  const Function<spacedim, typename InputVector::value_type> *>
     &                       neumann_bc,
   const InputVector &       solution,
-  Vector<float> &           error,
+  Vector<double> &          error,
   const ComponentMask &     component_mask,
   const Function<spacedim> *coefficients,
   const unsigned int        n_threads,
@@ -137,7 +137,7 @@ KellyErrorEstimator<1, spacedim>::estimate(
                  const Function<spacedim, typename InputVector::value_type> *>
     &                                     neumann_bc,
   const std::vector<const InputVector *> &solutions,
-  std::vector<Vector<float> *> &          errors,
+  std::vector<Vector<double> *> &         errors,
   const ComponentMask &                   component_mask,
   const Function<spacedim> *              coefficients,
   const unsigned int                      n_threads,
@@ -172,7 +172,7 @@ KellyErrorEstimator<1, spacedim>::estimate(
                  const Function<spacedim, typename InputVector::value_type> *>
     &                       neumann_bc,
   const InputVector &       solution,
-  Vector<float> &           error,
+  Vector<double> &          error,
   const ComponentMask &     component_mask,
   const Function<spacedim> *coefficients,
   const unsigned int        n_threads,
@@ -182,7 +182,7 @@ KellyErrorEstimator<1, spacedim>::estimate(
 {
   // just pass on to the other function
   const std::vector<const InputVector *> solutions(1, &solution);
-  std::vector<Vector<float> *>           errors(1, &error);
+  std::vector<Vector<double> *>          errors(1, &error);
   estimate(mapping,
            dof_handler,
            quadrature,
@@ -209,7 +209,7 @@ KellyErrorEstimator<1, spacedim>::estimate(
                  const Function<spacedim, typename InputVector::value_type> *>
     &                       neumann_bc,
   const InputVector &       solution,
-  Vector<float> &           error,
+  Vector<double> &          error,
   const ComponentMask &     component_mask,
   const Function<spacedim> *coefficients,
   const unsigned int        n_threads,
@@ -243,7 +243,7 @@ KellyErrorEstimator<1, spacedim>::estimate(
                  const Function<spacedim, typename InputVector::value_type> *>
     &                                     neumann_bc,
   const std::vector<const InputVector *> &solutions,
-  std::vector<Vector<float> *> &          errors,
+  std::vector<Vector<double> *> &         errors,
   const ComponentMask &                   component_mask,
   const Function<spacedim> *              coefficients,
   const unsigned int                      n_threads,
@@ -278,7 +278,7 @@ KellyErrorEstimator<1, spacedim>::estimate(
                  const Function<spacedim, typename InputVector::value_type> *>
     &                                     neumann_bc,
   const std::vector<const InputVector *> &solutions,
-  std::vector<Vector<float> *> &          errors,
+  std::vector<Vector<double> *> &         errors,
   const ComponentMask &                   component_mask,
   const Function<spacedim> *              coefficient,
   const unsigned int,
@@ -536,7 +536,7 @@ KellyErrorEstimator<1, spacedim>::estimate(
                  const Function<spacedim, typename InputVector::value_type> *>
     &                                     neumann_bc,
   const std::vector<const InputVector *> &solutions,
-  std::vector<Vector<float> *> &          errors,
+  std::vector<Vector<double> *> &         errors,
   const ComponentMask &                   component_mask,
   const Function<spacedim> *              coefficients,
   const unsigned int                      n_threads,
